pyproject.toml
[tool.poetry]
authors = ["Sandeep Aggarwal <asandeep.me@gmail.com>"]
description = "A sample electronics store management system"
name = "pseudo-electronics"
version = "0.1.0"

[tool.poetry.dependencies]
django = "^3.1.7"
django-crispy-forms = "^1.11.1"
django-role-permissions = "^3.1.1"
python = "^3.6"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2"
isort = "^5.7.0"
pre-commit = "2.1.1"
pylint = "^2.7.1"
pytest = "^5.2"
pytest-cov = "^2.11.1"


[tool.black]
line-length = 80


[tool.isort]
include_trailing_comma = true
line_length = 80
multi_line_output = 3


[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
